third_party/WebKit/LayoutTests/typedcssom/stylevalue-normalization/normalize-ident.html

<!doctype html>
<meta charset="utf-8">
<title>Identifier normalization tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#normalize-ident">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<script>
'use strict';

test(() => {
  assert_style_value_equals(CSSStyleValue.parse('width', 'auto'),
      new CSSKeywordValue('auto'));
}, 'CSS identifiers are normalized from String to CSSKeywordValues');

test(() => {
  assert_style_value_equals(
      newDivWithStyle('width: auto').styleMap.get('width'),
      new CSSKeywordValue('auto'));
}, 'CSS identifiers are normalized from CSSOM to CSSKeywordValues');

</script>
